+++ lib/src/deriv_chart/interactive_layer/interactive_layer.dart
@@ -15,6 +15,10 @@ import '../chart/data_visualization/drawing_tools/ray/ray_line_drawing.dart';
 import '../drawing_tool_chart/drawing_tools.dart';
 import 'interactable_drawing.dart';
 import 'interactable_drawing_custom_painter.dart';
+import 'interactive_layer_base.dart';
+import 'interactive_states/interactive_adding_tool_state.dart';
+import 'interactive_states/interactive_normal_state.dart';
+import 'interactive_states/interactive_state.dart';
 // ignore_for_file: public_member_api_docs
 
 /// Interactive layer of the chart package where elements can be drawn and can
@@ -312,190 +316,4 @@ class _InteractiveLayerGestureHandlerState
   void onAddDrawing(InteractableDrawing<dynamic> drawing) {
     widget.onAddDrawing?.call(drawing);
   }
-}
-
-abstract class InteractiveState {
-  InteractiveState({required this.interactiveLayer});
-
-  DrawingToolState getToolState(InteractableDrawing<DrawingToolConfig> drawing);
-
-  /// Additional drawings of the state to be drawn on top of the main drawings.
-  List<InteractableDrawing<DrawingToolConfig>> get additionalDrawings => [];
-
-  final InteractiveLayerBase interactiveLayer;
-
-  EpochFromX get epochFromX => interactiveLayer.epochFromX;
-
-  QuoteFromY get quoteFromY => interactiveLayer.quoteFromY;
-
-  EpochToX get epochToX => interactiveLayer.epochToX;
-
-  QuoteToY get quoteToY => interactiveLayer.quoteToY;
-
-  void onTap(TapUpDetails details);
-
-  void onPanUpdate(DragUpdateDetails details);
-
-  void onPanEnd(DragEndDetails details);
-
-  void onPanStart(DragStartDetails details);
-}
-
-class InteractiveNormalState extends InteractiveState {
-  InteractiveNormalState({required super.interactiveLayer});
-
-  @override
-  DrawingToolState getToolState(
-    InteractableDrawing<DrawingToolConfig> drawing,
-  ) =>
-      DrawingToolState.normal;
-
-  @override
-  void onPanEnd(DragEndDetails details) {}
-
-  @override
-  void onPanStart(DragStartDetails details) {}
-
-  @override
-  void onPanUpdate(DragUpdateDetails details) {}
-
-  @override
-  void onTap(TapUpDetails details) {
-    for (final drawing in interactiveLayer.drawings) {
-      if (drawing.hitTest(
-        details.localPosition,
-        epochToX,
-        quoteToY,
-      )) {
-        interactiveLayer.updateStateTo(
-          InteractiveSelectedToolState(
-            selected: drawing,
-            interactiveLayer: interactiveLayer,
-          ),
-        );
-        return;
-      }
-    }
-  }
-}
-
-class InteractiveSelectedToolState extends InteractiveState {
-  InteractiveSelectedToolState({
-    required this.selected,
-    required super.interactiveLayer,
-  });
-
-  final InteractableDrawing selected;
-
-  @override
-  DrawingToolState getToolState(
-      InteractableDrawing<DrawingToolConfig> drawing) {
-    return drawing.config.configId == selected.config.configId
-        ? DrawingToolState.selected
-        : DrawingToolState.normal;
-  }
-
-  @override
-  void onPanEnd(DragEndDetails details) {
-    selected.onDragEnd(details, epochFromX, quoteFromY, epochToX, quoteToY);
-  }
-
-  @override
-  void onPanStart(DragStartDetails details) {
-    selected.onDragStart(details, epochFromX, quoteFromY, epochToX, quoteToY);
-  }
-
-  @override
-  void onPanUpdate(DragUpdateDetails details) {
-    selected.onDragUpdate(details, epochFromX, quoteFromY, epochToX, quoteToY);
-  }
-
-  @override
-  void onTap(TapUpDetails details) {
-    for (final drawing in interactiveLayer.drawings) {
-      if (drawing.hitTest(
-        details.localPosition,
-        epochToX,
-        quoteToY,
-      )) {
-        interactiveLayer.updateStateTo(
-          InteractiveSelectedToolState(
-            selected: drawing,
-            interactiveLayer: interactiveLayer,
-          ),
-        );
-
-        return;
-      }
-    }
-
-    interactiveLayer.updateStateTo(
-      InteractiveNormalState(interactiveLayer: interactiveLayer),
-    );
-  }
-}
-
-class InteractiveAddingToolState extends InteractiveState {
-  InteractiveAddingToolState(
-    this.addingTool, {
-    required super.interactiveLayer,
-  });
-
-  final DrawingToolConfig addingTool;
-
-  InteractableDrawing<DrawingToolConfig>? _addingDrawing;
-
-  @override
-  List<InteractableDrawing<DrawingToolConfig>> get additionalDrawings =>
-      [if (_addingDrawing != null) _addingDrawing!];
-
-  @override
-  DrawingToolState getToolState(
-    InteractableDrawing<DrawingToolConfig> drawing,
-  ) =>
-      drawing.config.configId == addingTool.configId
-          ? DrawingToolState.adding
-          : DrawingToolState.normal;
-
-  @override
-  void onPanEnd(DragEndDetails details) {}
-
-  @override
-  void onPanStart(DragStartDetails details) {}
-
-  @override
-  void onPanUpdate(DragUpdateDetails details) {}
-
-  @override
-  void onTap(TapUpDetails details) {
-    _addingDrawing ??= addingTool.getInteractableDrawing();
-
-    _addingDrawing!
-        .onCreateTap(details, epochFromX, quoteFromY, epochToX, quoteToY, () {
-      interactiveLayer
-        ..clearAddingDrawing()
-        ..onAddDrawing(_addingDrawing!)
-        ..updateStateTo(
-          InteractiveNormalState(interactiveLayer: interactiveLayer),
-        );
-    });
-  }
-}
-
-abstract class InteractiveLayerBase {
-  void updateStateTo(InteractiveState state);
-
-  List<InteractableDrawing<DrawingToolConfig>> get drawings;
-
-  EpochFromX get epochFromX;
-
-  QuoteFromY get quoteFromY;
-
-  EpochToX get epochToX;
-
-  QuoteToY get quoteToY;
-
-  void clearAddingDrawing();
-
-  void onAddDrawing(InteractableDrawing<dynamic> drawing);
 }
